+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Fixed merchant columns sometimes resetting
 - Fixed populating items in the merchant not applying custom categories
+- Fixed opening the item pile UI of an unlinked token preventing opening the actor's default item pile UI
 
 ## Version 2.7.12 Hotfix